Add support for binary savestate as BSON, but turn it off and make binary savestates a stupid text clone because BSON sucks

+++ trunk/ExternalCoreProjects/Virtu/Services/VideoService.cs
@@ -16,10 +16,10 @@
 
 		public int[] fb;
 
-		public void SetPixel(int x, int y, uint color)
+		public void SetPixel(int x, int y, int color)
 		{
 			int i = 560 * y + x;
-			fb[i] = fb[i + 560] = (int)color;
+			fb[i] = fb[i + 560] = color;
 		}
     }
 }
